requirements.txt
pandas
numpy
nltk
scikit-learn
joblib
python-dotenv
pytest
pyyaml

